requirements.txt
asgiref==3.4.1
boto3==1.20.45
botocore==1.23.45
dj-database-url==0.5.0
Django==3.2.7
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.1
python-http-client==3.3.3
pytz==2021.1
s3transfer==0.5.0
sendgrid==6.1.0
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
